portfoliobackend/requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.14.0
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg==3.1.8
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
whitenoise==6.4.0
